+++ libs/angular/src/v-angular/table/table.mock.ts
@@ -73,23 +73,27 @@ export const columns: {
     {
       property: 'name',
       label: 'header.name',
+      hideLabelOnMobile: true,
+    },
+    {
+      property: 'status',
+      label: 'header.status',
+      hideLabelOnMobile: true,
     },
     {
       property: 'currency',
       label: 'header.currency',
     },
     {
-      property: 'status',
-      label: 'header.status',
-    },
-    {
       property: 'bookedBalance',
+      hidePropertyOnMobile: true,
       label: 'header.bookedbalance',
       valueType: 'numeric',
     },
     {
       property: 'datedBalance',
       label: 'header.datedbalance',
+      hidePropertyOnMobile: true,
       valueType: 'numeric',
     },
     {
@@ -99,6 +103,7 @@ export const columns: {
     },
     {
       property: 'unauthorizedUsage',
+      hidePropertyOnMobile: true,
       label: 'header.unauthorizedusage',
       valueType: 'numeric',
     },